Refactor: Contact layout
