+++ src/components/Portfolio/Philosophy/Philosophy.jsx
@@ -1,0 +1,12 @@
+import Title from '../Title/Title';
+import './Philosophy.scss';
+
+function Philosophy({ data }) {
+  return (
+    <div className="portfolio-inside-container">
+      <Title icon="ㅁ" />
+    </div>
+  );
+}
+
+export default Philosophy;
