+++ src/components/Portfolio/Portfolio.scss
@@ -3,6 +3,11 @@
   margin: 0 auto;
 }
 
+.portfolio-inside-container {
+  margin: 32px;
+  margin-top: 64px;
+}
+
 .portfolio-profile-bg-image {
   width: 100%;
   height: 360px;
