+++ src/components/Portfolio/Profile/Contact/Contact.scss
@@ -1,13 +1,13 @@
-.portfolio-contact-container {
-  margin: 32px;
-  margin-top: 64px;
-}
-
 .profile-contact-unit {
   display: flex;
   align-items: center;
   gap: 32px;
-  margin-top: 16px;
   text-decoration: none;
   color: black;
+}
+
+.portfolio-profile-contact {
+  margin-top: 8px;
+  display: flex;
+  gap: 8px;
 }
