+++ src/components/Portfolio/Profile/Profile.jsx
@@ -1,3 +1,4 @@
+import Contact from './Contact/Contact';
 import './Profile.scss';
 
 function Profile({ user }) {
@@ -14,6 +15,7 @@ function Profile({ user }) {
           안녕하세요! 네스트로 백엔드를 다루고, 플러터로 앱을 다루는 개발자
           김현우입니다. 인터넷 속에서는 `러리`라는 닉네임을 사용합니다.
         </div>
+        <Contact />
       </div>
     </div>
   );
